refactor(sync-server): home upload-chunk byte helpers in upload-size.ts
vault-deletion.ts needs parseUploadedChunks/getUploadedByteTotal to sum open
upload-session bytes. Homing them in services/blob.ts made routes/blob.ts import
them from a module that blob-upload-size-accounting.test.ts fully mocks, so the
route saw undefined helpers and returned 500. upload-size.ts is their real
domain and is not mocked; both call sites import from there.
